Create domain package
- Create domain package.
- Create job, process_step, job_status_type, step_type entity classes.
- Create exceptions package with JobNotFound exception class.
- Delete all unessesary modules from some packages.
