+++ src/jso_backend/domain/exceptions/job_not_found_exception.py
@@ -1,0 +1,4 @@
+class JobNotFoundError(Exception):
+    def __init__(self, *args: object, id: int) -> None:
+        super().__init__(*args)
+        self.message = f"Job with id {id} not found"
